Merge 0.6.9RC2 into master.

+++ package.py
@@ -48,7 +48,8 @@ def DeleteFiles(name):
     ''' Delete one file or set of files from wild-card spec '''
     for f in glob.glob(name):
         try:
-            os.remove(f)
+            if os.path.exists(f):
+                os.remove(f)
         except:
             print "Cannot remove file %s" % f
             exit(1)
@@ -70,26 +71,38 @@ def CheckPath(name):
 
 
 def PatchVersion(name):
-    """ Patch in the Bazaar baseline number, but only when this is
+    """ Patch in the Git commit hash, but only when this is
         an unmodified checkout
     """
     global my_version, my_baseline
 
+    commit = ''
     try:
-        pipe = subprocess.Popen(BzrVersion, shell=True, stdout=subprocess.PIPE).stdout
+        pipe = subprocess.Popen(GitVersion, shell=True, stdout=subprocess.PIPE).stdout
         for line in pipe.read().split('\n'):
-            if 'revno: ' in line:
-                bzr = line.split(' ')[1].strip()
+            if 'commit ' in line:
+                commit = line.split(' ')[1].strip()
+                break
         pipe.close()
     except:
-        pass
+        print 'Cannot run %s' % GitVersion
+        exit(1)
 
-    if not bzr:
-        print "WARNING: Cannot run %s" % BzrVersion
-        bzr = 'unknown'
+    state = ' (not committed)'
+    try:
+        pipe = subprocess.Popen(GitStatus, shell=True, stdout=subprocess.PIPE).stdout
+        for line in pipe.read().split('\n'):
+            if 'nothing to commit' in line:
+                state = ''
+                break
+        pipe.close()
+    except:
+        print 'Cannot run %s' % GitStatus
+        exit(1)
 
-    if not (bzr and bzr.isdigit()):
-        bzr = 'unknown'
+    if not commit:
+        print "WARNING: Cannot run %s" % GitVersion
+        commit = 'unknown'
 
     try:
         ver = open(VERSION_FILE, 'rb')
@@ -99,12 +112,12 @@ def PatchVersion(name):
         print "WARNING: cannot patch " + VERSION_FILE
         return
 
-    my_baseline = bzr
+    my_baseline = commit + state
     my_version = name
 
     regex = re.compile(r'__baseline__\s+=\s+"\w*"')
-    text = re.sub(r'__baseline__\s*=\s*"[^"]*"', '__baseline__ = "%s"' % bzr, text)
-    text = re.sub(r'__version__\s*=\s*"[^"]*"', '__version__ = "%s"' % name, text)
+    text = re.sub(r'__baseline__\s*=\s*"[^"]*"', '__baseline__ = "%s"' % my_baseline, text)
+    text = re.sub(r'__version__\s*=\s*"[^"]*"', '__version__ = "%s"' % my_version, text)
 
     try:
         ver = open(VERSION_FILE, 'wb')
@@ -120,14 +133,14 @@ def PairList(src):
         A dir returns for its root and each of its subdirs
             (path, <list-of-files>)
         Always return paths with Unix slashes.
-        Skip all Bazaar elements, .bak .pyc .pyo and *.~*
+        Skip all Git elements, .bak .pyc .pyo and *.~*
     """
     lst = []
     for item in src:
         if item.endswith('/'):
             for root, dirs, files in os.walk(item.rstrip('/\\')):
                 path = root.replace('\\', '/')
-                if path.find('.bzr') < 0:
+                if path.find('.git') < 0:
                     flist = []
                     for file in files:
                         if not (file.endswith('.bak') or file.endswith('.pyc') or file.endswith('.pyo') or '~' in file):
@@ -239,18 +252,27 @@ def check_runtimes():
         if path:
             path = os.path.join(path, 'Bazaar')
             if not os.path.exists(path):
-                path = None
-        if not path:
-            print 'Cannot find runtime libraries, have you installed Bazaar'
-            print 'in %s ?' % path
-            exit(1)
+                print 'Cannot find runtime libraries, have you installed Bazaar'
+                print 'in %s ?' % path
+                exit(1)
     return path
 
+def write_dll_message(path):
+    f = open(path, 'w')
+    f.write('''
+                          **** IMPORTANT ****
 
+If you get a Windows error message, claiming that DLL files are missing,
+please install "Microsoft Visual C++ 2008 Redistributable Package (x86)".
+Download from http://www.microsoft.com/download/en/details.aspx?displaylang=en&id=29
+Install this and *not* the SP1 package (or install both).
+
+''')
+    f.close()
 
 print sys.argv[0]
 
-Bazaar = CheckPath('bzr')
+Git = CheckPath('git')
 ZipCmd = CheckPath('zip')
 UnZipCmd = CheckPath('unzip')
 if os.name == 'nt':
@@ -258,12 +280,13 @@ if os.name == 'nt':
 else:
     NSIS = '-'
 
-BzrRevertApp =  Bazaar + ' revert --no-backup '
-BzrUpdateApp = Bazaar + ' update '
-BzrRevert =  Bazaar + ' revert --no-backup ' + VERSION_FILE
-BzrVersion = Bazaar + ' version-info'
+GitRevertApp =  Git + ' checkout -- '
+#GitUpdateApp = Git + ' update '
+GitRevertVersion =  GitRevertApp + ' ' + VERSION_FILE
+GitVersion = Git + ' log -1'
+GitStatus = Git + ' status'
 
-if not (Bazaar and ZipCmd and UnZipCmd and NSIS):
+if not (Git and ZipCmd and UnZipCmd and NSIS):
     exit(1)
 
 if len(sys.argv) < 2:
@@ -299,6 +322,7 @@ PatchVersion(release)
 
 # List of data elements, directories end with a '/'
 data_files = [
+         'ABOUT.txt',
          'README.txt',
          'INSTALL.txt',
          'GPL2.txt',
@@ -307,6 +331,7 @@ data_files = [
          'COPYRIGHT.txt',
          'ISSUES.txt',
          'nzb.ico',
+         'sabnzbd.ico',
          'Sample-PostProc.cmd',
          'Sample-PostProc.sh',
          'PKG-INFO',
@@ -341,18 +366,34 @@ options = dict(
 if target == 'app':
     if not platform.system() == 'Darwin':
         print "Sorry, only works on Apple OSX!"
-        os.system(BzrRevert)
+        os.system(GitRevertVersion)
         exit(1)
 
     # Check which Python flavour
     apple_py = 'ActiveState' not in sys.copyright
 
     #Create sparseimage from template
-    os.system("unzip sabnzbd-template.sparseimage.zip")
-    os.rename('sabnzbd-template.sparseimage', fileImg)
+    os.system("unzip osx/image/template.sparseimage.zip")
+    os.rename('template.sparseimage', fileImg)
 
     #mount sparseimage and modify volume label
     os.system("hdiutil mount %s | grep /Volumes/SABnzbd >mount.log" % (fileImg))
+
+    # Select OSX version specific background image
+    # Take care to preserve the special attributes of the background image file
+    if [int(n) for n in platform.mac_ver()[0].split('.')] >= [10, 7, 0]:
+        # Lion and higher
+        f = open('osx/image/sabnzbd_lion.png', 'rb')
+        png = f.read()
+        f.close()
+        f = open('/Volumes/SABnzbd/sabnzbd.png', 'wb')
+        f.write(png)
+        f.close()
+    else:
+        # Snow Leopard and lower
+        pass
+
+    # Rename the volume
     fp = open('mount.log', 'r')
     data = fp.read()
     fp.close()
@@ -407,13 +448,13 @@ if target == 'app':
     os.system("cp -pR osx/par2/ dist/SABnzbd.app/Contents/Resources/osx/par2>/dev/null")
     os.system("mkdir dist/SABnzbd.app/Contents/Resources/osx/unrar>/dev/null")
     os.system("cp -pR osx/unrar/ dist/SABnzbd.app/Contents/Resources/osx/unrar>/dev/null")
-    os.system("find dist/SABnzbd.app -name .bzr | xargs rm -rf")
+    os.system("find dist/SABnzbd.app -name .git | xargs rm -rf")
 
     #copy builded app to mounted sparseimage
     os.system("cp -r dist/SABnzbd.app /Volumes/%s/>/dev/null" % volume)
 
     print 'Create src %s' % fileOSr
-    os.system('tar -czf %s --exclude ".bzr*" --exclude "sab*.zip" --exclude "SAB*.tar.gz" --exclude "*.cmd" --exclude "*.pyc" '
+    os.system('tar -czf %s --exclude ".git*" --exclude "sab*.zip" --exclude "SAB*.tar.gz" --exclude "*.cmd" --exclude "*.pyc" '
               '--exclude "*.sparseimage" --exclude "dist" --exclude "build" --exclude "*.nsi" --exclude "win" --exclude "*.dmg" '
               './ >/dev/null' % (fileOSr) )
 
@@ -435,18 +476,17 @@ if target == 'app':
     os.system("hdiutil internet-enable %s" % fileDmg)
 
 
-    os.system(BzrRevertApp + "NSIS_Installer.nsi")
-    os.system(BzrRevertApp + VERSION_FILEAPP)
-    os.system(BzrRevertApp + VERSION_FILE)
-    os.system(BzrUpdateApp)
+    os.system(GitRevertApp + "NSIS_Installer.nsi")
+    os.system(GitRevertApp + VERSION_FILEAPP)
+    os.system(GitRevertApp + VERSION_FILE)
 
 elif target in ('binary', 'installer'):
     if not py2exe:
         print "Sorry, only works on Windows!"
-        os.system(BzrRevert)
+        os.system(GitRevertVersion)
         exit(1)
 
-    run_times = check_runtimes()
+    #run_times = check_runtimes()
 
     # Create MO files
     os.system('tools\\make_mo.py all')
@@ -522,13 +562,20 @@ elif target in ('binary', 'installer'):
 
 
     ############################
+    # Remove unwanted system DLL files that Py2Exe copies when running on Win7
+    DeleteFiles(r'dist\lib\API-MS-Win-*.dll')
+    DeleteFiles(r'dist\lib\MSWSOCK.DLL')
+    DeleteFiles(r'dist\lib\POWRPROF.DLL')
+
+    ############################
     # Copy MS runtime files or Curl
-    if run_times:
-        # MS Runtimes for Python 2.6+
-        shutil.copy2(os.path.join(run_times, r'Microsoft.VC90.CRT.manifest'), r'dist')
-        shutil.copy2(os.path.join(run_times, r'msvcp90.dll'), r'dist')
-        shutil.copy2(os.path.join(run_times, r'msvcr90.dll'), r'dist')
-        shutil.copy2(os.path.join(run_times, r'lib\Microsoft.VC90.CRT.manifest'), r'dist\lib')
+    if sys.version > (2, 5):
+        #Won't work with OpenSSL DLLs :(
+        #shutil.copy2(os.path.join(run_times, r'Microsoft.VC90.CRT.manifest'), r'dist')
+        #shutil.copy2(os.path.join(run_times, r'msvcp90.dll'), r'dist')
+        #shutil.copy2(os.path.join(run_times, r'msvcr90.dll'), r'dist')
+        #shutil.copy2(os.path.join(run_times, r'lib\Microsoft.VC90.CRT.manifest'), r'dist\lib')
+        pass
     else:
         # Curl for Python 2.5
         os.system(r'unzip -o win\curl\curl.zip -d dist\lib')
@@ -542,12 +589,13 @@ elif target in ('binary', 'installer'):
 
 
     DeleteFiles(fileBin)
+    write_dll_message('dist/IMPORTANT_MESSAGE.txt')
     os.rename('dist', prod)
     os.system('zip -9 -r -X %s %s' % (fileBin, prod))
     time.sleep(1.0)
     os.rename(prod, 'dist')
 
-    os.system(BzrRevert)
+    os.system(GitRevertVersion)
 
     ############################
     # Check for uncompressed sqlite3.dll
@@ -614,5 +662,5 @@ else:
     # Prepare the TAR.GZ pacakge
     CreateTar('srcdist', fileSrc, prod)
 
-    os.system(BzrRevert)
+    os.system(GitRevertVersion)
